feat(hook-error): rename
