+++ packages/falcon-data/src/Error/index.ts
@@ -1,2 +1,2 @@
-export * from './errorMapper';
+export * from './userError';
 export * from './useThrowError';
